Create README.md
I am slowly but surely getting the hang of exploring git. There is an order of operations I need to memorize on how to create/pull/clone/ or fork a repository, edit it as I need to, stage the commits, commit them, and then push them to the copied repository.